requirements.txt
pyyaml==5.4.1
pygithub==1.54.1
httpretty==1.0.5
